+++ browser-tests/pages/events.ts
@@ -2,4 +2,5 @@ import { Selector } from 'testcafe';
 
 export const eventsListPage = {
   listBody: Selector('.MuiTableBody-root'),
+  anyEventGroup: Selector('.MuiTableBody-root').withText('TAPAHTUMARYHMÄ'),
 };